requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# PDF Parsing
pymupdf==1.24.3
pdfplumber==0.11.0

# Embeddings
sentence-transformers==2.7.0

# Vector Store
faiss-cpu==1.8.0

# LLM
transformers==4.41.1
torch==2.3.0

# Data & Utils
numpy==1.26.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
loguru==0.7.2

# Frontend (optional)
streamlit==1.35.0
requests==2.32.2

sentencepiece==0.2.0
protobuf==4.25.3